recursive-coder/pyproject.toml
[project]
name = "recursive-coder"
version = "0.1.0"
description = "Recursive Task Decomposition AI Coding Framework"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
recursive-coder = "recursive_coder.cli:main"
